% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Rotations}
\alias{Rotations}
\title{Example of Rotations in package Tendril.}
\format{A numeric vector with 1000 number 3s:}
\usage{
Rotations
}
\description{
Example of Rotations in package Tendril.
}
\keyword{datasets}
