% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SubjList}
\alias{SubjList}
\title{Example of SubjList in package Tendril.}
\format{A data frame with 500 rows and 2 variables:
\describe{
  \item{subjid}{: The subject IDs}
  \item{treatment}{: The two treatments causing the tendrils to bend left or right}
}}
\usage{
SubjList
}
\description{
Example of SubjList in package Tendril.
}
\keyword{datasets}
