% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tendril.perm.res}
\alias{Tendril.perm.res}
\title{Example object Tendril as generated by Tendril() and Tendril.perm().}
\format{A list of 15 elements:
\describe{
\item{Tendril.perm.res$data }{: Dataframe of orginal data and coordinates}
\item{Tendril.perm.res$Terms }{: Column containing the name of the tendrils}
\item{Tendril.perm.res$Unique.Subject.Identifier }{: Columns containing subject IDs}
\item{Tendril.perm.res$Treat }{: Column containing the name of the treatments}
\item{Tendril.perm.res$StartDay }{: Column containing the start days of the events}
\item{Tendril.perm.res$Treatments }{: The names of the treatments causing the tendrils to bend}
\item{Tendril.perm.res$AEfreqThreshold }{: The frequency threshold for the events to be included in the analysis}
\item{Tendril.perm.res$Tag }{: A tag or comment associated with the analysis}
\item{Tendril.perm.res$SubjList }{: A dataframe containing subject IDs and treatments}
\item{Tendril.perm.res$SubjList.subject }{: The name of the column in SubjList containing the subjects IDs}
\item{Tendril.perm.res$SubjList.treatment }{: The name of the columns in SubjList containing the treatments}
\item{Tendril.perm.res$n.tot }{: A data frame of 2 variables for number of subjects in each treatment}
\item{Tendril.perm.res$PermTerm }{: Name of the tendril used for the permutations}
\item{Tendril.perm.res$perm.data }{: A dataframe containing the permutation results}
\item{Tendril.perm.res$tendril.pi }{: A list of permutation percentile calculation results}
}}
\usage{
Tendril.perm.res
}
\description{
The object contains the following 15 items:
}
\keyword{datasets}
