% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tendril.res}
\alias{Tendril.res}
\title{Example object Tendril as generated by Tendril().}
\format{A list of 12 elements:
\describe{
\item{Tendril.res$data }{: Dataframe of orginal data and coordinates}
\item{Tendril.res$Terms }{: Column containing the name of the tendrils}
\item{Tendril.res$Unique.Subject.Identifier }{: Columns containing subject IDs}
\item{Tendril.res$Treat }{: Column containing the name of the treatments}
\item{Tendril.res$StartDay }{: Column containing the start days of the events}
\item{Tendril.res$Treatments }{: The names of the treatments causing the tendrils to bend}
\item{Tendril.res$AEfreqThreshold }{: The frequency threshold for the events to be included in the analysis}
\item{Tendril.res$Tag }{: A tag or comment associated with the analysis}
\item{Tendril.res$SubjList }{: A dataframe containing subject IDs and treatments}
\item{Tendril.res$SubjList.subject }{: The name of the column in SubjList containing the subjects IDs}
\item{Tendril.res$SubjList.treatment }{: The name of the columns in SubjList containing the treatments}
\item{Tendril.res$SubjList.dropoutday }{: The name of the column in SubjList containing the dropoutday}
\item{Tendril.res$rotation_vector }{: The rotation vector used for the tendril analysis}
\item{Tendril.res$compensate_imbalance }{: Boolean indicating whether the imbalance in the groups is compensated}
\item{Tendril.res$n.tot }{: A data frame of 2 variables for number of subjects in each treatment}
}}
\usage{
Tendril.res
}
\description{
The object contains the following 12 items:
}
\keyword{datasets}
