% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TendrilData}
\alias{TendrilData}
\title{Example dataframe in package Tendril.}
\format{A data frame with 1000 rows and 4 variables:
\describe{
  \item{subjid}{The subject IDs}
  \item{treatment}{The two treatments causing the tendrils to bend right or left, respectively}
  \item{ae}{The names of the tendrils}
  \item{day}{The day of the event relative to the treatment start date}
}}
\usage{
TendrilData
}
\description{
A dataset containing the minimal 4 columns required to produe a Tendril plot
}
\keyword{datasets}
